Rename optimizely_decision.ts. to index.ts
