+++ packages/optimizely-sdk/lib/optimizely/index.ts
@@ -28,7 +28,7 @@ import {
   OptimizelyOptions,
   OptimizelyDecideOptions
 } from '../shared_types';
-import { OptimizelyDecision, newErrorDecision } from '../optimizely_decision/optimizely_decision';
+import { OptimizelyDecision, newErrorDecision } from '../optimizely_decision';
 import OptimizelyUserContext from '../optimizely_user_context';
 import { createProjectConfigManager, ProjectConfigManager } from '../core/project_config/project_config_manager';
 import { createNotificationCenter, NotificationCenter } from '../core/notification_center';
